Trait substrate_wasmtime::WasmTy

pub trait WasmTy { }

A trait implemented for types which can be arguments to closures passed to [Func::wrap1] and friends.

This trait should not be implemented by user types. This trait may change at any time internally. The types which implement this trait, however, are stable over time.

For more information see [Func::wrap1]